+++ src/Adapter/Element/ELEMENT.mjs
@@ -2,6 +2,8 @@ export const ELEMENT_CHOICE_SUBJECT = "choice-subject";
 
 export const ELEMENT_CREATE = "create";
 
+export const ELEMENT_INTENDED_DEGREE_PROGRAM = "intended-degree-program";
+
 export const ELEMENT_RESUME = "resume";
 
 export const ELEMENT_ROOT = "";
@@ -9,5 +11,5 @@ export const ELEMENT_ROOT = "";
 export const ELEMENT_START = "start";
 
 /**
- * @typedef {ELEMENT_CHOICE_SUBJECT | ELEMENT_RESUME | ELEMENT_CREATE | ELEMENT_ROOT | ELEMENT_START} ELEMENT
+ * @typedef {ELEMENT_CHOICE_SUBJECT | ELEMENT_CREATE | ELEMENT_INTENDED_DEGREE_PROGRAM | ELEMENT_RESUME | ELEMENT_ROOT | ELEMENT_START} ELEMENT
  */