+++ src/Adapter/IntendedDegreeProgram/ChosenIntendedDegreeProgram.mjs
@@ -1,0 +1,3 @@
+/**
+ * @typedef {{combination: string, subject: string}} ChosenIntendedDegreeProgram
+ */