+++ src/Adapter/IntendedDegreeProgram/continueFunction.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("./ChosenIntendedDegreeProgram.mjs").ChosenIntendedDegreeProgram} ChosenIntendedDegreeProgram */
+
+/**
+ * @typedef {(chosen_intended_degree_program: ChosenIntendedDegreeProgram) => void} continueFunction
+ */