+++ src/Adapter/Subject/Subject.mjs
@@ -1,0 +1,3 @@
+/**
+ * @typedef {{combinations: {id: string, label: string, mandatory?: string}[], id: string, label: string}} Subject
+ */